Create README.md
Library Management System is a system which maintains the information about the books present in the library, their authors, their publications and
their editions, etc. This is very difficult to organize manually. So basically the main objective of this project is to reduce the workload of management
as most of the manual work done is reduced.